Added Upload + Download DB menu items
I've added an upload database and download database menu item (currently not used and only in the debug menu). I've also moved the database file elements in the set database file dialog out of the panel and onto the dialog, along with starting on fields for controlling the dropbox data in the program outside of debug mode.