Renamed DATABASE_FILE_SOURCE
I've renamed the DATABASE_FILE_SOURCE constants to be EXTERNAL_DATABASE_FILE_SOURCE instead, since they will be used to indicate where the database file is stored externally in addition to locally. Additionally DATABASE_FILE_SOURCE_LOCAL has been changed to EXTERNAL_DATABASE_FILE_SOURCE_NONE. This is because the original intent was for the file only be stored in one location and not have a local copy if the database, but now the plan is to have a local and an optional external copy.