Removed encryption (for now)
I've removed the code and constants for encrypting and decrypting the dropbox access and refresh tokens. This is expected to be temporary, and once I've worked out more of the details, it will be reimplemented. It currently doesn't matter too much right now since I'm the only one who's going to be using it right now.