Fixed the autosave
The autosave was starting when the program creates a new file. This issue does not show up on small files, but on larger ones it causes the program to try to save it twice. Turns out it was only stopping the timer and not disabling the autosave frequency menu, which would tell the autosave timer it cannot start. I've also added a check for whether files are being saved to the function called by the timer to ensure that no files are being saved before the program starts autosaving the database.