Refined adding links to LinkMap
I've refined the process which the observed version of LinkMap.addAllSQL uses to add the links. I've made it so that it gets a map of all the longest prefixes and then inserts them using one of the insertSQL methods. This has resulted in a massive speed improvement (from ~45 minutes to ~15 seconds for approximately 100,000 new links) I've also made it so that the progress bar does not go indeterminate when commiting the changes to the database after a set number of links controlled by LINK_ADDING_AUTO_COMMIT.