Models are sorted before being saved
I've made it so that the LinksListModel that are being saved are sorted before they get saved, and any models that have the same listID will result all but one of the models with that listID having their listID removed (Assuming they aren't the same model). This should fix the possibility of a new database file assigning a listID currently used by another model to a different model before it adds the other model (thus it considers the other model already added to the database since it already has a model with that listID). This should also fix the possibility of two or more different models having the same listID.