Attempts at optimizing removed links
I'm attempting to optimize the code for removing unused links from the links table. I've tried collecting them all and removing them one by one, and currently I've got it so that it goes through and removes lists of them. This seems to have removed half a millisecond, but I'm not satisfied with that yet. I'll likely continue to work on it later on.