Updated the setContents method
As GLADoS once said, the best solution to the problem is usually the easiest one. As it turned out, the code for the setContents method was more complicated than it needed to be, as just clearing the list and adding the values from the given list was more efficient in the end. Perhaps the code used made sense and may have been more performant at some point, but not any more.