The remove unused rows code now has a default
I've added an abstract method to AbstractQueryRowMap which returns the name of the table that references the data in the map. I've also made it so that it has a default implementation of removeUnusedRowsSQL to AbstractQueryRowMap that removes any rows that are unused by the table that references the data. The implementations of PrefixMap, LinkMap, and ListNameMap in LinkDatabaseConnection have been updated accordingly.