External file path uses it's own property key
I've made it so that the path for the external (not local) copy of the database file uses a separate key, so as to ensure that, if the private configuration file is imported, it doesn't overwrite the path to the local database file. I've also added code to account for the older version of the private configuration and update it accordingly. I've also fixed an issue where a SQLiteConfig property that's dependent on the UI was being set before the UI was initialized.