Shapes are now protected and static
I've made it so that the shapes that do exists are now protected and static so that they can be used by multiple instances. This is because there could be many instances of these icons, so it would be a good idea to reduce the amount data used by each instance.