Added 3 Preference Nodes
I've added 3 currently unused preference nodes to the LinkManagerConfig. One of these is for the preference node for the program as a whole, one is for the preference node for instances of the program that share this instance's ID (I'll work on the ID soon), and one to store the sensitive data in a child of the local node.