Implemented the add and remove listener methods
I've implemented the methods for adding and removing PreferenceChangeListeners and NodeChangeListeners. These methods check if the node exists, and if so, adds or removes the listeners from the listener list. I've also made it so that the methods for adding listeners ignore nulls instead of throwing an exception, and the methods for removing listeners will ignore listeners that aren't registered to the node.