Made the typed put methods call the put method
(This was technically started in the last commit by accident) I made it so that the putInt, putLong, putBoolean, putFloat, and putDouble methods invoke the put(String, String) method, converting their values into Strings beforehand. This makes it so that if there are changes to the put method, it will apply to the other methods.