Removed a preference node + more
I've removed the privateNode preference node and renamed privateLocalNode back to privateNode. This is because the non-local privateNode was technically unnecessary. I've also renamed the PRIVATE_PREFERENCE_NODE_NAME constant to PRIVATE_PREFERENCE_NODE_PATH and added a "/" at the end so it can be used as a prefix for the private preference node. I've added a similar constant for the local preference node.