LinkManager now uses the new settings methods
I've made it so that LinkManager now uses the new methods in LinkManagerConfig for getting and setting the properties. Since most of the new methods for getting the properties now take in a default value, many of them are using the current value as the default value. I've also moved where the config has the hidden lists shown setting into the save config worker.