LinkManager now uses the importProperties
I've made it so that LinkManager now uses the importProperties method of the LinkManagerConfig. I've made it so that it's temporarily importing the normal properties, so that anything that is moved to the preference node will be carried over. This will not be left in the final version. I've also made ConfigLoader use the loadedConfig Properties object that was commented out.