Using the new property methods during import
I've made it so that the importProperties method is now using the new properties methods I just added. Also, I realized that checking for whether the properties map has a given key does not necessarily mean that it won't have the value, so instead I just get the property and deal with whether it's null or not.