put methods now accept nulls
I've made it so that the put and putByteArray (and by extension the putDimension, putPoint, and putRectangle methods) now accept null values, which remove the value currently set for the key. I've also added some documentation to the Dimension, Point, and Rectangle methods, which now mention the get, getByteArray, putByteArray, and the put or get method that corresponds to the method in question.