Added a putObject method
I added a putObject method which takes in an Object value and determines the appropriate put method to use to put the value into the preference node, removing the value if it's null, and defaulting to setting it as a String if no corresponding method is found. This is equivalent to the setPreference method in LinkManagerConfig.