Added header shorts
I've added some shorts to be used as headers to signify that a byte array may be a dimension, a point, or a rectangle. These won't guarantee that a byte array does not represent one of those, but it, along with a size check, should indicate that a byte array could be interpreted as one.