ByteBuffers used in private methods
I just realized ByteBuffers were a thing. I've created two private methods, one that converts an array of integers along with a header short into an array of bytes, and one that checks if a byte array starts with a header, and if so, returns the byte array as an IntBuffer (otherwise returning null).