Added prefixed property keys arrays and method
I've added two constant arrays storing the names of the configuration keys that serve as prefixes when in the Properties object. One of these is for the ones that deal with list types and the other deals with listIDs, with the latter currently being empty. I've also added a method to check if a given key matches one of these keys with their Properties suffix attached.