Added unfinished import code for remaining list keys
I've added the maps to get the values from the Properties object that is being imported, along with the code to add the properties to those maps and remove nulls at the end. I have not added the code to get the key prefixes for these maps so they are functionally unused, nor have I added the code putting the values in these maps into the preference nodes.