Current tab index map will never have null values
Given that the current tab index map is a view, and that view interprets null values as having no value for a key, then the map view will never have null values. As such, I removed the code that removes the null values from the map view when loading the settings.