Current tab and selected link update if fully loaded
The current tab's values and the selected links for the lists are only updated if the program has fully loaded. This should prevent a situation where the program hasn't fully loaded, and so the value is initialized to its default, which is accidentally stored into the preference node before the settings are loaded in.