Added a method for updating values ocasionally
I've added a method for updating the values in the program configuration that would be too difficult to update in real time or would update too frequently if updated in real time. I've already moved the search text and entered link into this method. Currently this method is only called when preparing the save configuration.