LinkManager now uses the preferences for settings
I've now made it so that LinkManager is now using the preferences for everything (except technically the programID which will stay in the configuration). This means that I now just need to do a bit of cleaning up and work on the programID stuff and a few other things to make this change complete.