Hidden lists are hidden at startup
I've made it so that, if the configuration has it set to show hidden lists but the lists are hidden after a set period of time, then the value for whether hidden lists are shown will be cleared when the program starts instead of being set to false when the program is closing. I've also fixed a glitch where the hidden list link operations enable value was being set to the general link operations enable value.