Added an AbstractConfigSaver class
I've added an abstract version of ConfigSaver which implements everything other than the actual saving, which is left up to an abstract method. This class also makes sure that the configuration is up to date. I've made ConfigSaver implement this class. I've also added a bit of documentation to the saveConfigFile method.