Added a ProgramConfigSaver worker
I've added a version of ConfigSaver called ProgramConfigSaver, which is what is now used to save the program's configuration file when the program is closing. ConfigSaver has had the constructors that take in the exit value and the constructors that don't take in a file removed since the ConfigSaver worker is going to just be used for saving the configuration to a file when the user asks to do so.