Search dialog location is stored
Another little side tangent, but I've now made it so that the search dialog's location is also stored. I've also made it so that if the location for the search dialog and the bounds for the set database location dialog aren't set, then they are made to be located relative to the program. I've also made the removal of the set database location dialog's size value from the config only occur when the program starts.