DB update helper methods use progress observer
I've made it so that the loadListDataIDs, repopulateListData, and repopulateListData methods now use a ProgressObserver to observe their progress instead of a JProgressBar. I've also temporarily added a method to create a ProgressObserver from a JProgressBar to help in the process of making the database update methods use ProgressObservers to observe their progress. This method, the toProgressObserver method, will be removed once it is no longer needed.