ProgressObserver now observes database updates
I've made it so that the program will now use a ProgressObserver to observe the progress when updating the database from database version 1.0.0 and beyond. The toProgressObserver method has also been removed, since it was just being used as a way to make the updated methods compatible with the ones that had yet to be updated.