Added new abstract database saver swing worker
I've added a new abstract database saving swing worker that will be used to replace the old one. This one incorporates functionality from the DatabaseFileSaver which will soon extend AbstractDatabaseFileSaver. I've also added an additional SavingStage called VERIFY_DATABASE for when the program is verifying the changes to the database.