Added more constructors to database file savers
I've added two additional constructors to AbstractDatabaseFileSaver and two to DatabaseSaver that don't take in a config file, instead relying on the getConfigFile method to get the config file. I've also added a check for whether a null config file was provided and, if so, the program won't try to save the configuration to a null config file.