Logging discrepancies in model sets
I've made it so that the program now logs if there's a discrepancy between the set created from the two tabs panels and the one maintained by the program. This will allow me to keep an eye on if there's ever an issue before I eventually use the set of models maintained by the program instead of this method.