Reworked DropboxAccountData to fit requirements
I've reworked DropboxAccountData to take in the FullAccount and SpaceUsage when given, and to derive them from the DbxUserUsersRequests if that is given instead. This allows me to remove the exceptions being thrown by the getter methods in order to fit within the AccountData interface specifications. Instead of throwing exceptions in the getter methods, the constructors may throw exceptions instead.